All subjects

+++ Core/ExamFunctions/Lvl2/max/max.c
@@ -6,7 +6,7 @@
 /*   By: gcarvalh <gcarvalh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/03 11:05:55 by gcarvalh          #+#    #+#             */
-/*   Updated: 2025/02/03 11:21:58 by gcarvalh         ###   ########.fr       */
+/*   Updated: 2025/02/19 07:03:36 by gcarvalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 int		max(int* tab, unsigned int len)
 {
 	int	max = tab[0];
-	int i = 0;
-	while (i < (int) len)
+	unsigned int i = 0;
+	while (i < len)
 	{
 		if (tab[i] > max)
 			max = tab[i];
